requirements.txt
fastybird-exchange-plugin>=0.4
fastybird-modules-metadata>=0.7
kink>=0.6
setuptools>=57.4
